Associate records before save
We ran into a bug due to a change in the way records are being
associated. When you are associating records in a singular association,
loading the record in a `before_create` callback will cause it to be
nil, and it will remain loaded as `nil` through the rest of the callback
flow.

In our case, we want to call a method on the belongs_to association during a
before_create callback that needs a reference to the association right away.
Using `build` instead of `build_record` will associate the two records together
before we save.